+++ psutil/_pslinux.py
@@ -21,7 +21,7 @@ from collections import namedtuple, defaultdict
 
 from psutil import _common
 from psutil import _psposix
-from psutil._common import (isfile_strict, usage_percent, deprecated)
+from psutil._common import isfile_strict, usage_percent, deprecated
 from psutil._compat import PY3
 import _psutil_linux as cext
 import _psutil_posix
@@ -291,7 +291,7 @@ def users():
         # to use them in the future.
         if not user_process:
             continue
-        if hostname == ':0.0':
+        if hostname == ':0.0' or hostname == ':0':
             hostname = 'localhost'
         nt = _common.suser(user, tty or None, hostname, tstamp)
         retlist.append(nt)
@@ -504,7 +504,7 @@ class Connections:
                 return []
         else:
             inodes = self.get_all_inodes()
-        ret = []
+        ret = set()
         for f, family, type_ in self.tmap[kind]:
             if family in (socket.AF_INET, socket.AF_INET6):
                 ls = self.process_inet(
@@ -519,8 +519,8 @@ class Connections:
                 else:
                     conn = _common.sconn(fd, family, type_, laddr, raddr,
                                          status, bound_pid)
-                ret.append(conn)
-        return ret
+                ret.add(conn)
+        return list(ret)
 
 
 _connections = Connections()