+++ psutil/_psosx.py
@@ -15,6 +15,7 @@ from collections import namedtuple
 from psutil import _common
 from psutil import _psposix
 from psutil._common import conn_tmap, usage_percent, isfile_strict
+from psutil._common import sockfam_to_enum, socktype_to_enum
 import _psutil_osx as cext
 import _psutil_posix
 
@@ -24,7 +25,7 @@ __extra__all__ = []
 # --- constants
 
 PAGESIZE = os.sysconf("SC_PAGE_SIZE")
-AF_LINK = socket.AF_LINK
+AF_LINK = _psutil_posix.AF_LINK
 
 # http://students.mimuw.edu.pl/lxr/source/include/net/tcp_states.h
 TCP_STATUSES = {
@@ -295,6 +296,8 @@ class Process(object):
         for item in rawlist:
             fd, fam, type, laddr, raddr, status = item
             status = TCP_STATUSES[status]
+            fam = sockfam_to_enum(fam)
+            type = socktype_to_enum(type)
             nt = _common.pconn(fd, fam, type, laddr, raddr, status)
             ret.append(nt)
         return ret