+++ psutil/_pssunos.py
@@ -15,7 +15,8 @@ from collections import namedtuple
 
 from psutil import _common
 from psutil import _psposix
-from psutil._common import usage_percent, isfile_strict
+from psutil._common import isfile_strict, socktype_to_enum, sockfam_to_enum
+from psutil._common import usage_percent
 from psutil._compat import PY3
 import _psutil_posix
 import _psutil_sunos as cext
@@ -212,7 +213,7 @@ def net_connections(kind, _pid=-1):
                          % (kind, ', '.join([repr(x) for x in cmap])))
     families, types = _common.conn_tmap[kind]
     rawlist = cext.net_connections(_pid, families, types)
-    ret = []
+    ret = set()
     for item in rawlist:
         fd, fam, type_, laddr, raddr, status, pid = item
         if fam not in families:
@@ -220,12 +221,14 @@ def net_connections(kind, _pid=-1):
         if type_ not in types:
             continue
         status = TCP_STATUSES[status]
+        fam = sockfam_to_enum(fam)
+        type_ = socktype_to_enum(type_)
         if _pid == -1:
             nt = _common.sconn(fd, fam, type_, laddr, raddr, status, pid)
         else:
             nt = _common.pconn(fd, fam, type_, laddr, raddr, status)
-        ret.append(nt)
-    return ret
+        ret.add(nt)
+    return list(ret)
 
 
 def wrap_exceptions(fun):